Remove ridiculous hard-coding of cancel method
so that it will use the one that is configured, but, madness and insanity,
the form to choose cancellation method is appearing despite permissions
set so that only true admin can, very much madness… only thing i didn't
test is if role hieararchy could be to blame.  Chased down every other
possible issue, it is not an overridden form.

#1308